Revert "Add new sv reward weight form  (#1733)"
Reverts #1733

Fixes https://github.com/DACH-NY/cn-test-failures/issues/5298 (which I don't understand)

[ci]
