--- apps/sv/frontend/src/__tests__/governance/forms/update-sv-reward-weight-form-test.test.tsx
@@ -1,220 +1,0 @@
-// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { fireEvent, render, screen, waitFor } from '@testing-library/react';
-import { describe, expect, test } from 'vitest';
-import { UpdateSvRewardWeightForm } from '../../../components/forms/UpdateSvRewardWeightForm';
-import userEvent from '@testing-library/user-event';
-import { SvConfigProvider } from '../../../utils';
-import App from '../../../App';
-import { svPartyId } from '../../mocks/constants';
-import { Wrapper } from '../../helpers';
-import { dateTimeFormatISO } from '@lfdecentralizedtrust/splice-common-frontend-utils';
-import dayjs from 'dayjs';
-
-describe('SV user can', () => {
-  test('login and see the SV party ID', async () => {
-    const user = userEvent.setup();
-    render(
-      <SvConfigProvider>
-        <App />
-      </SvConfigProvider>
-    );
-
-    expect(await screen.findByText('Log In')).toBeDefined();
-
-    const input = screen.getByRole('textbox');
-    await user.type(input, 'sv1');
-
-    const button = screen.getByRole('button', { name: 'Log In' });
-    user.click(button);
-
-    expect(await screen.findAllByDisplayValue(svPartyId)).toBeDefined();
-  });
-});
-
-describe('Update SV Reward Weight Form', () => {
-  test('should render all Update SV Reward Weight Form components', () => {
-    render(
-      <Wrapper>
-        <UpdateSvRewardWeightForm onSubmit={() => Promise.resolve()} />
-      </Wrapper>
-    );
-
-    expect(screen.getByTestId('update-sv-reward-weight-form')).toBeDefined();
-    expect(screen.getByText('Action')).toBeDefined();
-
-    const actionInput = screen.getByTestId('update-sv-reward-weight-action');
-    expect(actionInput).toBeDefined();
-    expect(actionInput.getAttribute('value')).toBe('SRARC_UpdateSvRewardWeight');
-
-    const summaryInput = screen.getByTestId('update-sv-reward-weight-summary');
-    expect(summaryInput).toBeDefined();
-    expect(summaryInput.getAttribute('value')).toBeNull();
-
-    const urlInput = screen.getByTestId('update-sv-reward-weight-url');
-    expect(urlInput).toBeDefined();
-    expect(urlInput.getAttribute('value')).toBe('');
-
-    const memberInput = screen.getByTestId('update-sv-reward-weight-member-dropdown');
-    expect(memberInput).toBeDefined();
-    expect(memberInput.getAttribute('value')).toBe('');
-
-    const weightInput = screen.getByTestId('update-sv-reward-weight-weight');
-    expect(weightInput).toBeDefined();
-    expect(weightInput.getAttribute('value')).toBe('');
-  });
-
-  test('should render errors when submit button is clicked on new form', async () => {
-    const user = userEvent.setup();
-
-    render(
-      <Wrapper>
-        <UpdateSvRewardWeightForm onSubmit={() => Promise.resolve()} />
-      </Wrapper>
-    );
-
-    const actionInput = screen.getByTestId('update-sv-reward-weight-action');
-    const submitButton = screen.getByTestId('submit-button');
-    expect(submitButton).toBeDefined();
-
-    await user.click(submitButton);
-    expect(submitButton.getAttribute('disabled')).toBeDefined();
-    expect(async () => await user.click(submitButton)).rejects.toThrowError(
-      /Unable to perform pointer interaction/
-    );
-
-    screen.getByText('Summary is required');
-    screen.getByText('Invalid URL');
-    screen.getByText('Weight is required');
-    screen.getByText('SV is required');
-
-    // completing the form should reenable the submit button
-    const summaryInput = screen.getByTestId('update-sv-reward-weight-summary');
-    expect(summaryInput).toBeDefined();
-    user.type(summaryInput, 'Summary of the proposal');
-
-    const urlInput = screen.getByTestId('update-sv-reward-weight-url');
-    expect(urlInput).toBeDefined();
-    user.type(urlInput, 'https://example.com');
-
-    const memberDropdown = screen.getByTestId('update-sv-reward-weight-member-dropdown');
-    expect(memberDropdown).toBeDefined();
-
-    const selectInput = screen.getByRole('combobox');
-    fireEvent.mouseDown(selectInput);
-
-    await waitFor(async () => {
-      const memberToSelect = screen.getByText('Digital-Asset-Eng-2');
-      expect(memberToSelect).toBeDefined();
-      await user.click(memberToSelect);
-    });
-
-    const weightInput = screen.getByTestId('update-sv-reward-weight-weight');
-    expect(weightInput).toBeDefined();
-    await user.type(weightInput, '1000');
-
-    await user.click(actionInput); // using this to trigger the onBlur event which triggers the validation
-
-    expect(submitButton.getAttribute('disabled')).toBe('');
-  });
-
-  test('expiry date must be in the future', async () => {
-    const user = userEvent.setup();
-    render(
-      <Wrapper>
-        <UpdateSvRewardWeightForm onSubmit={() => Promise.resolve()} />
-      </Wrapper>
-    );
-
-    const expiryDateInput = screen.getByTestId('update-sv-reward-weight-expiry-date-field');
-    expect(expiryDateInput).toBeDefined();
-
-    const thePast = dayjs().subtract(1, 'day').format(dateTimeFormatISO);
-    const theFuture = dayjs().add(1, 'day').format(dateTimeFormatISO);
-
-    await user.clear(expiryDateInput);
-    await user.type(expiryDateInput, thePast);
-
-    expect(screen.getByText('Expiration must be in the future')).toBeDefined();
-
-    await user.clear(expiryDateInput);
-    await user.type(expiryDateInput, theFuture);
-
-    expect(screen.queryByText('Expiration must be in the future')).toBeNull();
-  });
-
-  test('effective date must be after expiry date', async () => {
-    const user = userEvent.setup();
-
-    render(
-      <Wrapper>
-        <UpdateSvRewardWeightForm onSubmit={() => Promise.resolve()} />
-      </Wrapper>
-    );
-
-    const expiryDateInput = screen.getByTestId('update-sv-reward-weight-expiry-date-field');
-    const effectiveDateInput = screen.getByTestId('update-sv-reward-weight-effective-date-field');
-
-    const expiryDate = dayjs().add(1, 'week');
-    const effectiveDate = expiryDate.subtract(1, 'day');
-
-    await user.clear(expiryDateInput);
-    await user.type(expiryDateInput, expiryDate.format(dateTimeFormatISO));
-
-    await user.clear(effectiveDateInput);
-    await user.type(effectiveDateInput, effectiveDate.format(dateTimeFormatISO));
-
-    waitFor(() => {
-      expect(screen.getByText('Effective Date must be after expiration date')).toBeDefined();
-    });
-
-    const validEffectiveDate = expiryDate.add(1, 'day').format(dateTimeFormatISO);
-
-    await user.clear(effectiveDateInput);
-    await user.type(effectiveDateInput, validEffectiveDate);
-
-    expect(screen.queryByText('Effective Date must be after expiration date')).toBeNull();
-  });
-
-  test('sv dropdown contains all svs', async () => {
-    render(
-      <Wrapper>
-        <UpdateSvRewardWeightForm onSubmit={() => Promise.resolve()} />
-      </Wrapper>
-    );
-
-    const memberDropdown = screen.getByTestId('update-sv-reward-weight-member-dropdown');
-    expect(memberDropdown).toBeDefined();
-
-    const selectInput = screen.getByRole('combobox');
-    fireEvent.mouseDown(selectInput);
-
-    const svOptions = await screen.findAllByRole('option');
-    expect(svOptions.length).toBeGreaterThan(1);
-    expect(svOptions.map(option => option.textContent)).toEqual(
-      expect.arrayContaining(['Digital-Asset-2', 'Digital-Asset-Eng-2'])
-    );
-  });
-
-  test('Weight must be a valid number', async () => {
-    const user = userEvent.setup();
-    render(
-      <Wrapper>
-        <UpdateSvRewardWeightForm onSubmit={() => Promise.resolve()} />
-      </Wrapper>
-    );
-
-    const weightInput = screen.getByTestId('update-sv-reward-weight-weight');
-    expect(weightInput).toBeDefined();
-    await user.type(weightInput, '123abc');
-
-    expect(screen.getByText('Weight must be a valid number')).toBeDefined();
-
-    await user.clear(weightInput);
-    await user.type(weightInput, '1001');
-    await user.click(screen.getByTestId('update-sv-reward-weight-action'));
-
-    expect(screen.queryByText('Weight must be a valid number')).toBeNull();
-  });
-});
