+++ apps/sv/frontend/src/components/form-components/SelectField.tsx
@@ -1,54 +1,36 @@
 // Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-  Box,
-  FormControl,
-  FormHelperText,
-  MenuItem,
-  Select,
-  SelectChangeEvent,
-  Typography,
-} from '@mui/material';
+import { Box, FormControl, MenuItem, Select, SelectChangeEvent, Typography } from '@mui/material';
 import { useFieldContext } from '../../hooks/formContext';
 
 export type Option = { key: string; value: string };
 export interface SelectFieldProps {
   title: string;
   options: Option[];
-  id: string;
 }
 
 export const SelectField: React.FC<SelectFieldProps> = props => {
-  const { title, options, id } = props;
+  const { title, options } = props;
   const field = useFieldContext<string>();
-
   return (
-    <Box data-testid={`${id}-select-component`}>
+    <Box>
       <Typography variant="h6" gutterBottom>
         {title}
       </Typography>
-
-      <FormControl variant="outlined" error={!field.state.meta.isValid} fullWidth>
+      <FormControl variant="outlined" fullWidth>
         <Select
           value={field.state.value}
           onChange={(e: SelectChangeEvent) => field.handleChange(e.target.value as string)}
           onBlur={field.handleBlur}
-          error={!field.state.meta.isValid}
-          id={`${id}-dropdown`}
-          inputProps={{ 'data-testid': `${id}-dropdown` }}
         >
-          {options.map((member, index) => (
-            <MenuItem
-              key={'option-' + index}
-              value={member.value}
-              data-testid={`option-${member.key}`}
-            >
-              {member.key}
-            </MenuItem>
-          ))}
+          {options &&
+            options.map((member, index) => (
+              <MenuItem key={'option-' + index} value={member.key}>
+                {member.value}
+              </MenuItem>
+            ))}
         </Select>
-        <FormHelperText>{field.state.meta.errors?.[0]}</FormHelperText>
       </FormControl>
     </Box>
   );
