+++ apps/sv/frontend/src/components/forms/CreateProposalform.tsx
@@ -2,7 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { useSearchParams } from 'react-router-dom';
-import { Box, FormControlLabel, Radio, RadioGroup, TextField, Typography } from '@mui/material';
+import {
+  Box,
+  Button,
+  FormControlLabel,
+  Paper,
+  Radio,
+  RadioGroup,
+  TextField,
+  Typography,
+} from '@mui/material';
 import dayjs from 'dayjs';
 import { dateTimeFormatISO } from '@lfdecentralizedtrust/splice-common-frontend-utils';
 import { useMemo, useState } from 'react';
@@ -10,7 +19,6 @@ import { createProposalActions } from '../../utils/governance';
 import { useDsoInfos } from '../../contexts/SvContext';
 import { buildDsoConfigChanges } from '../../utils/buildDsoConfigChanges';
 import { useAppForm } from '../../hooks/form';
-import { FormLayout } from './FormLayout';
 
 type ConfigFormData = Record<string, string>;
 
@@ -18,14 +26,6 @@ type CreateProposalAction = (typeof createProposalActions)[number];
 
 interface CreateProposalFormProps {
   action: CreateProposalAction;
-}
-
-export interface CommonFormData {
-  action: string;
-  expiryDate: string;
-  effectiveDate: string;
-  url: string;
-  summary: string;
 }
 
 type FormValues = {
@@ -46,8 +46,8 @@ export const CreateProposalForm: React.FC<CreateProposalFormProps> = ({ action }
   const dsoInfoQuery = useDsoInfos();
 
   const memberOptions: { key: string; value: string }[] = [
-    { key: 'Super Validator 1', value: 'sv1' },
-    { key: 'Super Validator 2', value: 'sv2' },
+    { key: 'sv1', value: 'Super Validator 1' },
+    { key: 'sv2', value: 'Super Validator 2' },
   ];
 
   const defaultValues = useMemo((): FormValues => {
@@ -106,115 +106,157 @@ export const CreateProposalForm: React.FC<CreateProposalFormProps> = ({ action }
   const dsoConfigChanges = buildDsoConfigChanges(dsoConfig, dsoConfig, true);
 
   return (
-    <FormLayout form={form} id="create-proposal-form">
-      <Typography variant="h3">[Work In Progress]</Typography>
-
-      <form.Field
-        name="action"
-        children={field => {
-          return (
-            <Box>
-              <Typography variant="h5" gutterBottom>
-                Action
-              </Typography>
-              <TextField
-                fullWidth
-                variant="outlined"
-                autoComplete="off"
-                name={field.name}
-                value={field.state.value}
-                disabled
-              />
-            </Box>
-          );
-        }}
-      />
-
-      <form.AppField name="expiryDate">
-        {field => (
-          <field.DateField
-            title="Vote Proposal Expiration"
-            description="This is the last day voters can vote on this proposal"
-            id="expiry-date"
-          />
-        )}
-      </form.AppField>
-
-      <form.Field
-        name="effectiveDate"
-        children={_ => {
-          return (
-            <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-              <Typography variant="h5" gutterBottom>
-                Vote Proposal Effectivity
-              </Typography>
-
-              <RadioGroup
-                value={effectivityType}
-                onChange={e => setEffectivityType(e.target.value)}
-              >
-                <FormControlLabel
-                  value="custom"
-                  control={<Radio />}
-                  label={<Typography>Custom</Typography>}
+    <>
+      <Box sx={{ mt: 10 }}>
+        <Paper
+          sx={{
+            bgcolor: 'background.paper',
+            p: 4,
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+          }}
+        >
+          <Box sx={{ minWidth: '80%' }}>
+            <form
+              onSubmit={e => {
+                e.preventDefault();
+                e.stopPropagation();
+                form.handleSubmit();
+              }}
+            >
+              <Box sx={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
+                <Typography variant="h3">[Work In Progress]</Typography>
+                <form.Field
+                  name="action"
+                  children={field => {
+                    return (
+                      <Box>
+                        <Typography variant="h5" gutterBottom>
+                          Action
+                        </Typography>
+                        <TextField
+                          fullWidth
+                          variant="outlined"
+                          autoComplete="off"
+                          name={field.name}
+                          value={field.state.value}
+                          disabled
+                        />
+                      </Box>
+                    );
+                  }}
                 />
 
-                {effectivityType === 'custom' && (
-                  <form.AppField name="effectiveDate">
-                    {field => (
-                      <field.DateField
-                        description="Select the date and time the proposal will take effect"
-                        id="effective-date"
-                      />
+                <form.AppField name="expiryDate">
+                  {field => (
+                    <field.DateField
+                      title="Vote Proposal Expiration"
+                      description="This is the last day voters can vote on this proposal"
+                    />
+                  )}
+                </form.AppField>
+
+                <form.Field
+                  name="effectiveDate"
+                  children={_ => {
+                    return (
+                      <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
+                        <Typography variant="h5" gutterBottom>
+                          Vote Proposal Effectivity
+                        </Typography>
+
+                        <RadioGroup
+                          value={effectivityType}
+                          onChange={e => setEffectivityType(e.target.value)}
+                        >
+                          <FormControlLabel
+                            value="custom"
+                            control={<Radio />}
+                            label={<Typography>Custom</Typography>}
+                          />
+
+                          {effectivityType === 'custom' && (
+                            <form.AppField name="effectiveDate">
+                              {field => (
+                                <field.DateField description="Select the date and time the proposal will take effect" />
+                              )}
+                            </form.AppField>
+                          )}
+
+                          <FormControlLabel
+                            value="threshold"
+                            control={<Radio />}
+                            label={
+                              <Box>
+                                <Typography>Make effective at threshold</Typography>
+                                <Typography variant="body2" color="text.secondary">
+                                  This will allow the vote proposal to take effect immediately when
+                                  2/3 vote in favor
+                                </Typography>
+                              </Box>
+                            }
+                            sx={{ mt: 2 }}
+                          />
+                        </RadioGroup>
+                      </Box>
+                    );
+                  }}
+                />
+
+                <form.AppField name="summary">
+                  {field => <field.TextArea title="Proposal Summary" optional />}
+                </form.AppField>
+
+                <form.AppField name="url">{field => <field.TextField title="URL" />}</form.AppField>
+
+                <form.AppField name="member">
+                  {field => <field.SelectField title="Member" options={memberOptions} />}
+                </form.AppField>
+
+                <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
+                  <Typography variant="h6" gutterBottom>
+                    Configuration
+                  </Typography>
+
+                  {dsoConfigChanges.map((change, index) => (
+                    <form.AppField name={change.fieldName} key={index}>
+                      {field => <field.ConfigField configChange={change} key={index} />}
+                    </form.AppField>
+                  ))}
+                </Box>
+
+                <Box
+                  sx={{
+                    display: 'flex',
+                    justifyContent: 'center',
+                    alignItems: 'center',
+                    mt: 4,
+                    spacing: 4,
+                  }}
+                >
+                  <Button variant="outlined" sx={{ mr: 8 }}>
+                    Cancel
+                  </Button>
+                  <form.Subscribe
+                    selector={state => [state.canSubmit, state.isSubmitting]}
+                    children={([canSubmit, isSubmitting]) => (
+                      <Button
+                        variant="pill"
+                        type={'submit'}
+                        size="large"
+                        disabled={!canSubmit || isSubmitting}
+                      >
+                        {isSubmitting ? 'Submitting' : 'Submit Proposal'}
+                      </Button>
                     )}
-                  </form.AppField>
-                )}
-
-                <FormControlLabel
-                  value="threshold"
-                  control={<Radio />}
-                  label={
-                    <Box>
-                      <Typography>Make effective at threshold</Typography>
-                      <Typography variant="body2" color="text.secondary">
-                        This will allow the vote proposal to take effect immediately when 2/3 vote
-                        in favor
-                      </Typography>
-                    </Box>
-                  }
-                  sx={{ mt: 2 }}
-                />
-              </RadioGroup>
-            </Box>
-          );
-        }}
-      />
-
-      <form.AppField name="summary">
-        {field => <field.TextArea title="Proposal Summary" optional id="summary" />}
-      </form.AppField>
-
-      <form.AppField name="url">{field => <field.TextField title="URL" id="url" />}</form.AppField>
-
-      <form.AppField name="member">
-        {field => <field.SelectField title="Member" options={memberOptions} id="member" />}
-      </form.AppField>
-
-      <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-        <Typography variant="h6" gutterBottom>
-          Configuration
-        </Typography>
-
-        {dsoConfigChanges.map((change, index) => (
-          <form.AppField name={change.fieldName} key={index}>
-            {field => <field.ConfigField configChange={change} key={index} />}
-          </form.AppField>
-        ))}
+                  />
+                </Box>
+              </Box>
+            </form>
+          </Box>
+        </Paper>
       </Box>
-
-      <form.AppForm>
-        <form.FormControls />
-      </form.AppForm>
-    </FormLayout>
+    </>
   );
 };
