+++ apps/sv/frontend/src/hooks/form.ts
@@ -8,7 +8,6 @@ import { TextField } from '../components/form-components/TextField';
 import { TextArea } from '../components/form-components/TextArea';
 import { SelectField } from '../components/form-components/SelectField';
 import { ConfigField } from '../components/form-components/ConfigField';
-import { FormControls } from '../components/form-components/FormControls';
 
 export const { useAppForm } = createFormHook({
   fieldComponents: {
@@ -18,9 +17,7 @@ export const { useAppForm } = createFormHook({
     TextArea,
     TextField,
   },
-  formComponents: {
-    FormControls,
-  },
+  formComponents: {},
   fieldContext,
   formContext,
 });
